:truck: General renaming and file moving
Rename fill components and files to better represent them. Also created
a new package in order to better separate the concerns.
