+++ features/task/src/main/java/com/escodro/task/di/TaskModule.kt
@@ -4,7 +4,7 @@ import com.escodro.task.mapper.AlarmIntervalMapper
 import com.escodro.task.mapper.CategoryMapper
 import com.escodro.task.mapper.TaskMapper
 import com.escodro.task.mapper.TaskWithCategoryMapper
-import com.escodro.task.presentation.TaskSectionViewModel
+import com.escodro.task.presentation.list.TaskListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -14,7 +14,7 @@ import org.koin.dsl.module
 val taskModule = module {
 
     // Presentation
-    viewModel { TaskSectionViewModel(/*get(), get()*/) }
+    viewModel { TaskListViewModel(/*get(), get()*/) }
 
     // Mappers
     factory { AlarmIntervalMapper() }
