+++ features/task/src/main/java/com/escodro/task/presentation/list/TaskListViewState.kt
@@ -1,4 +1,4 @@
-package com.escodro.task.presentation
+package com.escodro.task.presentation.list
 
 import com.escodro.task.model.TaskWithCategory
